Change python3.8 --> python3
I really think it is a bad practice to hard code the python version.
For example on my system I don't have python3.8 (I am already at
python3.9) so it fails.

The idea is that python3 should be a pointer to whatever is the most
recent version of python3 on the current system. Since we already check
that it is aleat 3.8, hard coding the version in the Makefile is
unnecessary.