Use order-only prerequisites in Makefile, fix #11
This is important not only to guarantee idempotence of the linking step
but also to make sure no issues arises from parallel processing of the
Makefile (which might be enabled by default on some platform).

This is necessary since our linking procedure is inherently sequential:
if we are trying to match against a larger pool of officers, there are
more ambiguities which needs to be resolved by "hand".